Refactor mute handling and update script paths for improved organization
- Replaced references to `test_report_mute.py` with `mute_utils.py` in various scripts to align with the new structure.
- Introduced `get_muted_tests.py` to enhance functionality for managing muted tests.
- Updated import statements across multiple scripts to reflect the new module organization, improving maintainability.
- Removed the obsolete `path_resolver.py` file to streamline the codebase.
