+++ .github/workflows/update_muted_ya.yml
@@ -125,7 +125,7 @@ jobs:
 
       - name: Update muted tests in DB
         if: steps.mute_gate.outputs.skip != 'true'
-        run: python3 .github/scripts/tests/get_muted_tests.py upload_muted_tests --branch=${{ env.BASE_BRANCH }} --build_type=${{ env.BUILD_TYPE }} --muted_ya_file=base_muted_ya.txt
+        run: python3 .github/scripts/tests/mute/get_muted_tests.py upload_muted_tests --branch=${{ env.BASE_BRANCH }} --build_type=${{ env.BUILD_TYPE }} --muted_ya_file=base_muted_ya.txt
 
       - name: Update test monitor
         if: steps.mute_gate.outputs.skip != 'true'
